Simplify OAuth provider plumbing
