+++ plugins/_oauth/helpers/providers/gemini_api.py
@@ -5,7 +5,7 @@ import json
 import time
 from pathlib import Path
 from typing import Any
-from urllib.parse import parse_qs, urlencode, urlparse
+from urllib.parse import urlencode, urlparse
 
 from plugins._oauth.helpers.providers.base import (
     DUMMY_API_KEY,
@@ -19,12 +19,16 @@ from plugins._oauth.helpers.providers.base import (
     read_json_file,
     write_private_json,
 )
-from plugins._oauth.helpers.usage_plans import (
-    usage_plan_notes_for,
-    usage_plan_sources_for,
-    usage_plans_for,
+from plugins._oauth.helpers.providers.common import (
+    as_int as _as_int,
+    as_optional_string as _as_optional_string,
+    error_message as _error_message,
+    expires_ms as _expires_ms,
+    json_payload as _json_payload,
+    latest_attempt,
+    models_from_payload as _models_from_payload,
+    parse_manual_callback,
 )
-from plugins._oauth.helpers import state as state_store
 from plugins._oauth.helpers.state import get_attempt, pop_attempt, put_attempt
 
 
@@ -46,34 +50,6 @@ CLIENT_CONFIG_NOTE = (
     "This uses Gemini API billing/quotas, not Antigravity or Gemini Code Assist subscription quota."
 )
 REFRESH_MARGIN_MS = 60_000
-
-
-def parse_manual_callback(raw: Any) -> dict[str, str | None] | None:
-    text = "" if raw is None else str(raw).strip()
-    if not text:
-        return None
-
-    if text.startswith("http://") or text.startswith("https://"):
-        query = urlparse(text).query
-    elif text.startswith("?"):
-        query = text[1:]
-    elif "=" in text or "&" in text:
-        query = text
-    else:
-        return {
-            "code": text,
-            "state": None,
-            "error": None,
-            "error_description": None,
-        }
-
-    parsed = parse_qs(query, keep_blank_values=True)
-    return {
-        "code": _first_query_value(parsed, "code"),
-        "state": _first_query_value(parsed, "state"),
-        "error": _first_query_value(parsed, "error"),
-        "error_description": _first_query_value(parsed, "error_description"),
-    }
 
 
 class GeminiApiOAuthProvider:
@@ -106,9 +82,6 @@ class GeminiApiOAuthProvider:
             supports_oauth_client_config=True,
             supports_quota_project=True,
             note=CLIENT_CONFIG_NOTE,
-            usage_plans=usage_plans_for(GEMINI_API_PROVIDER_ID),
-            usage_plan_notes=usage_plan_notes_for(GEMINI_API_PROVIDER_ID),
-            usage_plan_sources=usage_plan_sources_for(GEMINI_API_PROVIDER_ID),
         )
 
     def status(self) -> dict[str, Any]:
@@ -302,7 +275,7 @@ class GeminiApiOAuthProvider:
         if state:
             attempt = get_attempt(state)
             if attempt is None:
-                if _latest_gemini_attempt() is not None:
+                if latest_attempt(GEMINI_API_PROVIDER_ID) is not None:
                     return LoginPollResult(
                         ok=False,
                         provider_id=GEMINI_API_PROVIDER_ID,
@@ -321,7 +294,7 @@ class GeminiApiOAuthProvider:
                     error="OAuth state mismatch. Return to Agent Zero and start a new Google Gemini API connection.",
                 )
         elif allow_missing_state:
-            attempt = _latest_gemini_attempt()
+            attempt = latest_attempt(GEMINI_API_PROVIDER_ID)
             if attempt is None:
                 return LoginPollResult(
                     ok=False,
@@ -642,78 +615,6 @@ def _validate_google_token_endpoint(value: str) -> None:
         )
 
 
-def _latest_gemini_attempt():
-    state_store.cleanup_expired()
-    with state_store._lock:
-        attempts = [
-            attempt
-            for attempt in state_store._attempts.values()
-            if attempt.provider_id == GEMINI_API_PROVIDER_ID and not attempt.expired()
-        ]
-    if not attempts:
-        return None
-    return max(attempts, key=lambda attempt: attempt.created_at)
-
-
-def _models_from_payload(payload: Any) -> list[str]:
-    values: list[Any]
-    if isinstance(payload, dict) and isinstance(payload.get("data"), list):
-        values = payload["data"]
-    elif isinstance(payload, dict) and isinstance(payload.get("models"), list):
-        values = payload["models"]
-    elif isinstance(payload, list):
-        values = payload
-    else:
-        return []
-
-    models: list[str] = []
-    seen: set[str] = set()
-    for value in values:
-        model_id = ""
-        if isinstance(value, str):
-            model_id = value
-        elif isinstance(value, dict):
-            model_id = str(value.get("id") or value.get("name") or "")
-        model_id = model_id.strip()
-        if model_id.startswith("models/"):
-            model_id = model_id.split("/", 1)[1]
-        if model_id and model_id not in seen:
-            seen.add(model_id)
-            models.append(model_id)
-    return models
-
-
-def _json_payload(response: Any) -> dict[str, Any]:
-    try:
-        payload = response.json()
-    except Exception:
-        payload = {}
-    if not isinstance(payload, dict):
-        return {}
-    return payload
-
-
-def _error_message(payload: dict[str, Any], fallback: str) -> str:
-    error = payload.get("error")
-    if isinstance(error, dict):
-        return str(error.get("message") or error.get("status") or fallback)
-    return str(payload.get("error_description") or payload.get("error") or fallback)
-
-
-def _first_query_value(parsed: dict[str, list[str]], key: str) -> str | None:
-    values = parsed.get(key) or []
-    if not values:
-        return None
-    return values[0]
-
-
-def _as_optional_string(value: Any) -> str | None:
-    if isinstance(value, list):
-        value = value[0] if value else None
-    text = "" if value is None else str(value).strip()
-    return text or None
-
-
 def _account_label(auth: dict[str, Any]) -> str:
     claims = _jwt_claims(str(auth.get("id_token") or ""))
     return str(claims.get("email") or auth.get("account_label") or "Google Gemini API")
@@ -730,28 +631,6 @@ def _jwt_claims(token: str) -> dict[str, Any]:
     except Exception:
         return {}
     return parsed if isinstance(parsed, dict) else {}
-
-
-def _expires_ms(payload: dict[str, Any]) -> int:
-    if payload.get("expires_at") is not None:
-        try:
-            value = float(payload["expires_at"])
-            if value < 1_000_000_000_000:
-                value *= 1000
-            return int(value)
-        except (TypeError, ValueError):
-            pass
-    try:
-        return int((time.time() + float(payload.get("expires_in") or 0)) * 1000)
-    except (TypeError, ValueError):
-        return 0
-
-
-def _as_int(value: Any, default: int) -> int:
-    try:
-        return int(value)
-    except (TypeError, ValueError):
-        return default
 
 
 def _codex_helper():
